Move the ImportLocaleCommand to the new Console namespace
